Pre-convert exact_cycle into bytes for buffering
Instead of using cycle and islice to repeat the individual bytes from
exact_cycle until the specified num_samples is met, convert exact_cycle
into a bytes object using b''.join() and repeat said bytes until the
specified num_samples is met, cutting off the last repetition if
necessary.

This provides around a 9 to 10x speedup in the code, with the only
bottleneck now being the actual write() call to the file.